virtual-client/requirements.txt
python-socketio==4.6.1
requests==2.25.1
